fix bugs

+++ app/src/main/java/com/example/android/databinding/basicsample/data/remote/response/movie/detail/MovieDetailResponse.kt
@@ -26,7 +26,7 @@ data class MovieDetailResponse(
         val genres: List<GenresItem>,
 
         @field:SerializedName("popularity")
-        val popularity: Double? = null,
+        val popularity: Float? = null,
 
         @field:SerializedName("production_countries")
         val productionCountries: List<ProductionCountriesItem?>? = null,
