+++ app/src/main/java/com/example/android/databinding/basicsample/data/remote/response/tvshow/detail/TvShowDetailResponse.kt
@@ -23,7 +23,7 @@ data class TvShowDetailResponse(
 	val genres: List<GenresItem>,
 
 	@field:SerializedName("popularity")
-	val popularity: Double? = null,
+	val popularity: Float? = null,
 
 	@field:SerializedName("id")
 	val id: Int? = null,
